# specs/001-terminal-designer-routing/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: LiquidTwin API (feature 001)
  version: 0.1.0
  description: Terminal management, validation, availability and quickest-route queries. No authentication in this feature.
servers:
  - url: /api/v1
paths:
...
  /terminals/{terminalId}/routes:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
    post:
      summary: Quickest feasible routes for one operation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteRequest' }
      responses:
        '200':
          description: Routes (possibly empty) with explanations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /terminals/{terminalId}/availability:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
    get:
      summary: List availability windows
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: element_id, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AvailabilityWindow' }
    post:
      summary: Upsert availability windows (idempotent by element_id+source+external_ref). Unknown elements are reported, not fatal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [windows]
              properties:
                windows:
                  type: array
                  maxItems: 5000
                  items: { $ref: '#/components/schemas/AvailabilityWindowInput' }
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                type: object
                required: [applied, rejected]
                properties:
                  applied: { type: integer }
                  rejected:
                    type: array
                    items:
                      type: object
                      required: [index, reason]
                      properties:
                        index: { type: integer }
                        reason: { type: string }
  /terminals/{terminalId}/availability/{windowId}:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
      - { name: windowId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Delete a window
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  parameters:
    TerminalId:
      name: terminalId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Input rejected; nothing was saved
      content:
        application/json:
          schema:
            type: object
            required: [issues]
            properties:
              issues:
                type: array
                items: { $ref: '#/components/schemas/ValidationIssue' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    ValidationIssue:
      type: object
      required: [severity, code, message]
      properties:
        severity: { type: string, enum: [ERROR, WARNING] }
        code: { type: string }
        element_id: { type: string }
        node_id: { type: string }
        row: { type: integer, description: CSV row for import errors }
        file: { type: string }
        message: { type: string }
        fix_hint: { type: string }
    AvailabilityWindowInput:
      type: object
      required: [element_id, status, from]
      properties:
        element_id: { type: string }
        status: { $ref: '#/components/schemas/AvailabilityStatus' }
        from: { type: string, format: date-time }
        to: { type: [string, 'null'], format: date-time }
        reason: { type: string }
        source: { type: string, default: manual }
        external_ref: { type: string }
    AvailabilityWindow:
      allOf:
        - { $ref: '#/components/schemas/AvailabilityWindowInput' }
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
    AvailabilityStatus:
      type: string
      enum: [AVAILABLE, MAINTENANCE, FLUSHING, CLEANING, OUT_OF_SERVICE]
    RouteRequest:
      type: object
      required: [direction, product_id, volume_m3, rate_m3h, window_from, source_ids, destination_ids]
      properties:
        version: { type: integer, minimum: 1, description: Defaults to latest }
        direction: { type: string, enum: [IN, OUT, TRANSFER] }
        product_id: { type: string }
        volume_m3: { type: number, exclusiveMinimum: 0 }
        rate_m3h: { type: number, exclusiveMinimum: 0 }
        window_from: { type: string, format: date-time }
        window_to: { type: string, format: date-time, description: Defaults to window_from plus transfer time }
        source_ids: { type: array, minItems: 1, items: { type: string } }
        destination_ids: { type: array, minItems: 1, items: { type: string } }
        max_routes: { type: integer, minimum: 1, maximum: 20, default: 5 }
        pump_suction_inputs:
          type: array
          items:
            type: object
            required: [pump_id, npsh_available_m]
            properties:
              pump_id: { type: string }
              npsh_available_m: { type: number, minimum: 0 }
...
    RouteResponse:
      type: object
      required: [terminal_version, engine_version, routes, exclusions]
      properties:
        terminal_version: { type: integer }
        engine_version: { type: string }
        routes:
          type: array
          items: { $ref: '#/components/schemas/Route' }
        exclusions:
          type: array
          items: { $ref: '#/components/schemas/Exclusion' }
        no_route:
          $ref: '#/components/schemas/NoRoute'
    Route:
      type: object
      required: [rank, source_id, destination_id, steps, metrics]
      properties:
        rank: { type: integer }
        source_id: { type: string }
        destination_id: { type: string }
        steps:
          type: array
          items:
            type: object
            required: [element_id, from, to, reversed]
            properties:
              element_id: { type: string }
              from: { type: string }
              to: { type: string }
              reversed: { type: boolean }
        metrics:
          type: object
          required: [fill_min, transfer_min, flush_min, flush_volume_m3, total_min, valves, common_headers, head_margin_m, max_velocity_ms]
          properties:
            fill_min: { type: number }
            transfer_min: { type: number }
            flush_min: { type: number }
            flush_volume_m3: { type: number }
            total_min: { type: number }
            valves: { type: integer }
            common_headers: { type: integer }
            head_margin_m: { type: number }
            max_velocity_ms: { type: number }
            operating_flow_m3h: { type: [number, 'null'] }
            pump_head_m: { type: [number, 'null'] }
            system_head_m: { type: [number, 'null'] }
            suction_margin_m: { type: [number, 'null'] }
            pump_speed_ratio: { type: [number, 'null'] }
    Exclusion:
      type: object
      required: [reason, detail]
      properties:
        element_id: { type: string }
        node_id: { type: string }
        reason:
          type: string
          enum: [NOT_CERTIFIED, VELOCITY, RESIDUE, ONE_WAY_PUMP, UNAVAILABLE, DEDICATED_OTHER_GROUP, PUMP_HEAD, ENDPOINT_STOCK, ENDPOINT_SPACE, ENDPOINT_PRODUCT, ENDPOINT_DIRECTION, INVALID_PUMP_CURVE, PUMP_FLOW_OUT_OF_RANGE, NO_PUMP_SYSTEM_INTERSECTION, PUMP_SPEED_OUT_OF_RANGE, PUMP_SUCTION_MARGIN]
        detail: { type: string }
...
    NoRoute:
      type: object
      required: [message, blockers]
      properties:
        message: { type: string }
        blockers:
          type: array
          items: { $ref: '#/components/schemas/Exclusion' }